Trait vortex_sampling_compressor::compressors::EncoderMetadata

pub trait EncoderMetadata {
    // Required method
    fn as_any(&self) -> &dyn Any;
}

Metadata that can optionally be attached to a compression tree.

This enables codecs to cache trained parameters from the sampling runs to reuse for the large run.

Required Methods

fn as_any(&self) -> &dyn Any